perfect!!

+++ include/angles.h
@@ -8,9 +8,9 @@
 #include "point.h"
 
 struct EulerAngles {
-    float pitch; // x-rotation
-    float yaw;   // y-rotation
-    float roll;  // z-rotation
+    double pitch; // x-rotation
+    double yaw;   // y-rotation
+    double roll;  // z-rotation
 };
 
 Matrix3d EulerToMatrix(const EulerAngles& angle);
@@ -18,13 +18,13 @@ Matrix3d EulerToMatrix(const EulerAngles& angle);
 Matrix3d EulerToMatrixLH(const EulerAngles& angle);
 
 
-void movePointForward(Point3& pointPosition, EulerAngles &angles, float distance);
+void movePointForward(Point3& pointPosition, EulerAngles &angles, double distance);
 
-void movePointRight(Point3& pointPosition, EulerAngles &angles, float distance);
+void movePointRight(Point3& pointPosition, EulerAngles &angles, double distance);
 
-void movePointBack(Point3& pointPosition, EulerAngles &angles, float distance);
+void movePointBack(Point3& pointPosition, EulerAngles &angles, double distance);
 
-void movePointLeft(Point3& pointPosition, EulerAngles &angles, float distance);
+void movePointLeft(Point3& pointPosition, EulerAngles &angles, double distance);
 
 
 #endif /* ANGLE_H */
