+++ include/point.h
@@ -7,9 +7,9 @@
 
 class Point3 {
 public:
-	float x, y, z;
+	double x, y, z;
 	Point3() : x(0), y(0), z(0) {}
-	Point3(float x, float y, float z) : x(x), y(y), z(z) {}
+	Point3(double x, double y, double z) : x(x), y(y), z(z) {}
 
 	Point3 operator-(const Point3& other) const;
 	Point3 operator+(const Point3& other) const;
@@ -17,9 +17,9 @@ public:
 
 class Point2 {
 public:
-	float x, y;
+	double x, y;
 	Point2() : x(0), y(0) {}
-	Point2(float x, float y) : x(x), y(y) {}
+	Point2(double x, double y) : x(x), y(y) {}
 };
 
 #endif /* POINT_H */
