+++ include/vector.h
@@ -8,14 +8,14 @@
 
 class Vector2 {
 public:
-	float x, y;
+	double x, y;
 
 	Vector2() : x(0), y(0) {}
-	Vector2(float x, float y) : x(x), y(y) {}
+	Vector2(double x, double y) : x(x), y(y) {}
 
-	float dot(const Vector2& vec) const;
+	double dot(const Vector2& vec) const;
 
-	float magnitude() const;
+	double magnitude() const;
 
 	Vector2 normalize() const;
 
@@ -23,21 +23,21 @@ public:
 
 	Vector2 operator-(const Vector2& vec) const;
 
-	Vector2 operator*(float scalar) const;
+	Vector2 operator*(double scalar) const;
 };
 
 class Vector3 {
 public:
-	float x, y, z;
+	double x, y, z;
 
 	Vector3() : x(0), y(0), z(0) {}
-	Vector3(float x, float y, float z) : x(x), y(y), z(z) {}
+	Vector3(double x, double y, double z) : x(x), y(y), z(z) {}
 
 	Vector3 cross(const Vector3 &vec) const;
 
-	float dot(const Vector3& vec) const;
+	double dot(const Vector3& vec) const;
 	
-	float magnitude() const;
+	double magnitude() const;
 
 	Vector3 normalize() const;
 
@@ -45,7 +45,7 @@ public:
 
 	Vector3 operator-(const Vector3& vec) const;
 
-	Vector3 operator*(float scalar) const;
+	Vector3 operator*(double scalar) const;
 
 	Point3 operator=(const Point3& other) const;
