+++ src/vector2.cpp
@@ -1,10 +1,10 @@
 #include "../include/vector.h"
 
-float Vector2::dot(const Vector2& vec) const {
+double Vector2::dot(const Vector2& vec) const {
     return (x * vec.x) + (y * vec.y);
 }
 
-float Vector2::magnitude() const {
+double Vector2::magnitude() const {
     return std::sqrt(x * x + y * y);
 }
 
@@ -22,7 +22,7 @@ Vector2 Vector2::operator-(const Vector2& vec) const {
     );
 }
 
-Vector2 Vector2::operator*(float scalar) const {
+Vector2 Vector2::operator*(double scalar) const {
     return Vector2(
         x * scalar,
         y * scalar
@@ -30,6 +30,6 @@ Vector2 Vector2::operator*(float scalar) const {
 }
 
 Vector2 Vector2::normalize() const {
-    float mag = magnitude();
+    double mag = magnitude();
     return (mag > 0) ? Vector2(x / mag, y / mag) : Vector2(0, 0);
 }
